app/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
google-cloud-storage==2.14.0
google-cloud-bigquery==3.20.0
